% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asgntax.R
\name{asgntax}
\alias{asgntax}
\title{Taxonomy assignment based on external reference database in BiMiCo pipeline}
\usage{
asgntax(asvtab, taxdat, revcomp, mtthread)
}
\arguments{
\item{asvtab}{(Required) ASV table object (of class matrix) created by the 'std_(454/Illum)_ASVtab' function (ASVs in rows, samples in columns)}

\item{taxdat}{(Required) Path to taxonomy database file compatible with DADA2 'assignTaxonomy' function}

\item{mtthread}{(Required) Boolean, enables multithreading (not recommended in Rstudio)}
}
\description{
Generates taxonomic assignment of ASV sequences, given an ASV table generated by BiMiCo and an external taxonomic reference database compatible with DADA2 'assignTaxonomy' function. Currently Greengenes v13.5 is provided with the package.
}
\examples{
asgntax()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
