% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_454_ASVtab.R
\name{asvtab_454}
\alias{asvtab_454}
\title{Create ASV table from filtered 454 reads}
\usage{
asvtab_454(readfiles, mtthread)
}
\arguments{
\item{readfiles}{(Required) Path to 454 quality-filtered fastq files}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio) (default=F)}
}
\description{
Given a set of 454 sequencing reads pre-filtered by the 'prep_454' function, internally generates read error models and outputs chimera-filtered ASV table. ASV table contains ASVs in rows and samples in columns.
}
\examples{
asvtab_454()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
