% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_illSE_ASVtab.R
\name{asvtab_illSE}
\alias{asvtab_illSE}
\title{Create ASV table from filtered SINGLE-END Illumina reads}
\usage{
asvtab_illSE(readfiles, mtthread)
}
\arguments{
\item{readfiles}{(Required) Path to SE Illumina quality-filtered fastq files}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio) Default=F}
}
\description{
Given a set of SE Illumina sequencing reads pre-filtered by the 'prep_illSE' function, internally generates read error models and outputs chimera-filtered ASV table. ASV table contains ASVs in rows and samples in columns.
}
\examples{
asvtab_illSE()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
