% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimico_454.R
\name{bimico_454}
\alias{bimico_454}
\title{single-step preprocessing of 454 reads in BiMiCo pipeline}
\usage{
bimico_454(readfiles, taxfile, pheno, outdir, trim_read_length = 0,
  mtthread = F)
}
\arguments{
\item{readfiles}{(Required) Path to single-end 454 demultiplexed fastq files, extension currently has to be ".fastq"}

\item{taxfile}{(Required) Path to taxonomy database}

\item{pheno}{(Required) Phenotye table}

\item{outdir}{(Required) Path to write quality-filtered fastq files to}

\item{trim_read_length}{(Optional) max. length at which to truncate reads. Default = 0 (no truncating)}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio). Default=F}
}
\description{
Read processing function. Returns quality-filtered reads in output folder and Phyloseq object of ASVs, taxonomy and phenodata.
}
\examples{
bimico_454()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
