% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_read_quality.R
\name{insp_reads}
\alias{insp_reads}
\title{inspect read trimming results}
\usage{
insp_reads(readfiles, filtfiles, filenum = 1)
}
\arguments{
\item{readfiles}{(Required) Path to single-end 454 demultiplexed fastq files}

\item{filtfiles}{(Required) Path to quality-filtered fastq files generated by std_(454/illum_SE/illum_PE)_preprocess}

\item{filenum}{(Optional) Number of fastq read file to plot (numeric).}
}
\description{
Convenience wrapper for DADA2 read processing functions. Generates QC plots from a given fastq file before and after quality control by DADA2.
}
\examples{
insp_reads()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
