% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiem_decontamination.R
\name{MicrobIEM_decontamination}
\alias{MicrobIEM_decontamination}
\title{Microbiem decontamination by Luise Rauer, TUM Munich}
\usage{
MicrobIEM_decontamination(feature_table, SAMPLE, NEG1 = character(0),
  NEG2 = character(0), ratio_NEG1_threshold = NA,
  ratio_NEG2_threshold = NA, span_NEG1_threshold = NA,
  span_NEG2_threshold = NA)
}
\arguments{
\item{feature_table}{(Required) data.frame, a feature table with count data, with ASVs/OTUs/taxa in rows, ASV IDs as rownames, and samples in columns}

\item{SAMPLE}{(Required) character vector, names of samples to be filtered}

\item{NEG1}{(Optional) character vector, names of NEG1 controls, can by empty}

\item{NEG2}{(Optional) character vector of names of NEG2 controls, can be empty}

\item{ratio_NEG1_threshold}{(Optional) Numeric, thresholds for filtering contaminants, default is to not use any filter (NA)}

\item{ratio_NEG2_threshold}{(Optional) Numeric, thresholds for filtering contaminants, default is to not use any filter (NA)}

\item{span_NEG1_threshold}{(Optional) Numeric, thresholds for filtering contaminants, default is to not use any filter (NA)}

\item{span_NEG2_threshold}{(Optional) Numeric, thresholds for filtering contaminants, default is to not use any filter (NA)}
}
\description{
Numeric thresholds for filtering contaminants, default is to not use any filter (NA)
ratio_NEG1_threshold = NA, # Sensible thresholds: > 0
ratio_NEG2_threshold = NA, # Sensible thresholds: > 0
span_NEG1_threshold = NA, # Sensible thresholds: between 0 and 1
span_NEG2_threshold = NA # Sensible thresholds: between 0 and 1
}
\examples{
MicrobIEM_decontamination()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
