% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCoA_phylo.R
\name{pcoa_ph}
\alias{pcoa_ph}
\title{Pricipal coordinate analysis (PCoA)}
\usage{
pcoa_ph(data, dotcol = "black")
}
\arguments{
\item{data}{(Required) The microbiota data on which you want to perform ordination needs to represent phyloseq-class.i.e. the data should include the components of otu_table_class, taxonomyTable_class and sample_data_class.otu_table_class should be organised so that rows represent samples and columns represent OTUs.}

\item{dotcol}{(Optional) column name in sample_data of the phyloseq object specified to argument "data". Uses phyloseq default color palette (default:black)}
}
\value{
Returns principal coordinate analysis object.
}
\description{
Principal coordinate analysis using sample-wise Bray-Curtis dissimilarity index.
}
\examples{
pcoa_ph()

}
