% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_illum_SE_preprocessor.R
\name{prep_illSE}
\alias{prep_illSE}
\title{standard preprocessor for SINGLE-END Illumina reads in BiMiCo pipeline}
\usage{
prep_illSE(readfiles, outdir, mtthread = F, trim_read_length = 0,
  trim5end = 0)
}
\arguments{
\item{readfiles}{(Required) Path to single-end Illumina demultiplexed fastq files}

\item{outdir}{(Required) String specifying output directory for processed files; will be created as subdirectory of input dir.}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio). Default=F}

\item{trim_read_length}{(Optional) max. length at which to truncate reads. Default = 0 (no truncating)}

\item{trim5end}{(Optional) trim n reads from the start of reads (unidentified adapters, barcodes). Default=0 (no trimming)}
}
\description{
Convenience wrapper for DADA2 read processing functions. Given a directory of raw Illumina fastq files, performs platform-specific trimming and QC, writes reads to output directory.
}
\examples{
prep_illSE()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
