% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_asvs.R
\name{rename_asvs}
\alias{rename_asvs}
\title{Convert Phyloseq object with ASVs renamed to more convenient shorter form, and actual ASV sequences stored as a DNAStringSet (refseq slot)}
\usage{
rename_asvs(phylo)
}
\arguments{
\item{phylo}{(Required) input object of class 'phyloseq'}
}
\description{
Takes as input a phyloseq object and converts otu_table and tax_table rownames to shorter form, stores ASV sequences in refseq slot as 'DNAStringSet'
}
\examples{
rename_asvs()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
