% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.nonbac.R
\name{rm_nonbac}
\alias{rm_nonbac}
\title{Removal of non-bacterial sequences from a phyloseq object}
\usage{
rm_nonbac(x)
}
\arguments{
\item{x}{Phyloseq object with non-empty otu_table and tax_table slots; otu_table may contain either OTUs or ASVs.}
}
\description{
Given an input phyloseq object with valid tax_table, returns taxonomy-filtered phyloseq object with mitochondrial, chloroplast and non-bacterial taxa removed.
Non-bacterial sequences (e.g. mitochondria, chloroplast) need to be removed from OTU/ASV tables prior to statistical analyses. rm_nonbac removes all rows of otu_table and tax_table in a phyloseq object based on taxonomic labelling 'Mitochondria', 'Chloroplast' or where Kingdom is other than 'Bacteria'. NAs in taxonomy table are converted to 'unidentified'.
}
\examples{
rm_nonbac()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
