% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_454_ASVtab.R
\name{std_454_ASVtab}
\alias{std_454_ASVtab}
\title{Create ASV table from filtered 454 reads}
\usage{
std_454_ASVtab(readfiles)
}
\arguments{
\item{readfiles}{(Required) Path to 454 quality-filtered fastq files}
}
\description{
Given a set of 454 sequencing reads pre-filtered by the 'std_454_preprocess' function, internally generates read error models and outputs chimera-filtered ASV table. ASV table contains ASVs in rows and samples in columns.
}
\examples{
std_454_ASVtab()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
