% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_454_preprocessor.R
\name{std_454_preprocess}
\alias{std_454_preprocess}
\title{standard preprocessor for 454 reads in BiMiCo pipeline}
\usage{
std_454_preprocess(readfiles, outdir, trim_read_length = 0)
}
\arguments{
\item{readfiles}{(Required) Path to single-end 454 demultiplexed fastq files}

\item{outdir}{(Required) String specifying output directory for processed files; will be created as subdirectory of input dir.}

\item{trim_read_length}{(Optional) max. length at which to truncate reads. Default = 0 (no truncating)}
}
\description{
Convenience wrapper for DADA2 read processing functions. Given a directory of raw 454 fastq files, performs platform-specific trimming and QC, outputs reads to a sub-folder of the specified name
}
\examples{
std_454_preprocess()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
