% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_ASV_tables.R
\name{tidy_ASV_tables}
\alias{tidy_ASV_tables}
\title{tidy_ASV_tables}
\usage{
tidy_ASV_tables(asvtab)
}
\arguments{
\item{asvtab}{(Required) ASV table generated by bimico pipeline}
}
\description{
Rename ASVs with more convenient short names
}
\examples{
tidy_ASV_tables()
}
\keyword{16S}
\keyword{analysis}
\keyword{data}
