% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvtab_454.R
\name{asvtab_454}
\alias{asvtab_454}
\title{Create ASV table from filtered 454 reads}
\usage{
asvtab_454(readfiles, n, mtthread = F, pattern = "_filt.fastq.gz",
  extname = "_filt.fastq.gz")
}
\arguments{
\item{readfiles}{(Required) Path to 454 quality-filtered fastq files}

\item{n}{(Optional) The maximum number of reads to dereplicate at any one time. Controls peak memory requirement. Default=1e6}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio) (default=F)}

\item{pattern}{(Optional) Pattern (e.g., extension) in input fastq filename, only files with this pattern will be read (default="_filt.fastq.gz")}

\item{extname}{(Optional) Suffix of input files. (default="_filt.fastq.gz")}
}
\description{
Given a set of 454 sequencing reads pre-filtered by the 'prep_454' function, internally generates read error models and outputs chimera-filtered ASV table. ASV table contains ASVs in rows and samples in columns.
}
\examples{
asvtab_454()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
