% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvtab_illPE.R
\name{asvtab_illPE}
\alias{asvtab_illPE}
\title{Create ASV table from filtered PAIRED-END Illumina reads}
\usage{
asvtab_illPE(fwd_reads, rev_reads, filt_out, rawext, mtthread = F,
  mergepairs = F)
}
\arguments{
\item{fwd_reads}{(Required) Path to PE Illumina quality-filtered fastq files}

\item{rev_reads}{(Required) Path to PE Illumina quality-filtered fastq files}

\item{filt_out}{(Required) Path to filtered reads from previous step (TrimandFilter)}

\item{rawext}{(Required) Raw extension, as in previous steps}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio) Default=F}

\item{mergepairs}{(Optional) Boolean, to merge read pairs (recommended) or only concatenate (if there's no overlap between the majority of reads, it is recommended to inspect trimming). Default=F}
}
\description{
Given a set of PE Illumina sequencing reads pre-filtered by the 'prep_illSE' function, internally generates read error models and outputs chimera-filtered ASV table. ASV table contains ASVs in rows and samples in columns.
}
\examples{
asvtab_illPE()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
