% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvtab_illSE_v1.R
\name{asvtab_illSE}
\alias{asvtab_illSE}
\title{Create ASV table from filtered SINGLE-END Illumina reads}
\usage{
asvtab_illSE(readfiles, batch, n = 1e+06, mtthread = F)
}
\arguments{
\item{readfiles}{(Required) Path to SE Illumina quality-filtered fastq DIRECTORY (e.g. filt_fqs_dir in tutorial)}

\item{batch}{(Required) List indicating the sequencing batch (e.g., a column of phenodata with batch information)}

\item{n}{(Optional) The maximum number of reads to dereplicate at any one time. Controls peak memory requirement. Default=1e6}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio) Default=F}
}
\description{
Given a set of SE Illumina sequencing reads pre-filtered by the 'prep_illSE' function, internally generates read error models and outputs chimera-filtered ASV table. ASV table contains ASVs in rows and samples in columns.
}
\examples{
asvtab_illSE()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
