% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimico_illPE.R
\name{bimico_illPE}
\alias{bimico_illPE}
\title{single-step preprocessing of 454 reads in BiMiCo pipeline}
\usage{
bimico_illPE(fwd_reads, rev_reads, taxfile, pheno, filt_out, outdir,
  trim_read_length = 0, mtthread = F)
}
\arguments{
\item{taxfile}{(Required) Path to taxonomy database}

\item{pheno}{(Required) Phenotye table}

\item{outdir}{(Required) Path to write quality-filtered fastq files to}

\item{trim_read_length}{(Optional) max. length at which to truncate reads. Default = 0 (no truncating)}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio). Default=F}

\item{readfiles}{(Required) Path to single-end 454 demultiplexed fastq files, extension currently has to be ".fastq"}
}
\description{
Read processing function. Returns quality-filtered reads in output folder and Phyloseq object of ASVs, taxonomy and phenodata.
}
\examples{
bimico_454()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
