% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phylo.R
\name{create_phylo}
\alias{create_phylo}
\title{Summarize ASV table and taxonomy in Phyloseq object, with optional phenotype data}
\usage{
create_phylo(asvtb, taxtb, phenodata)
}
\arguments{
\item{phenodata}{(Optional) sample phenotype matrix, if available; sample names must match with ASV table sample names.}

\item{asvtab}{(Required) ASV table generated in BimiCo (ASVs in rows)}

\item{taxtab}{(Required) Taxonomy table generated in BiMiCo}
}
\description{
Takes as input an ASV table and taxonomy table created in the BiMiCo pipeline, with optional sample phenotypes imported e.g. from a csv table, and creates a Phyloseq object.
}
\examples{
create_phylo()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
