% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_454.R
\name{prep_454}
\alias{prep_454}
\title{standard preprocessor for 454 sequencing reads in BiMiCo pipeline}
\usage{
prep_454(readfiles, outdir, file_suffix = ".fastq", mtthread = F,
  trim_read_length = 0, trim5end = 0)
}
\arguments{
\item{readfiles}{(Required) Path to 454 demultiplexed fastq files}

\item{outdir}{(Required) String specifying output directory for processed files.}

\item{file_suffix}{(Optional) Please specify if input files have special suffixes other than ".fastq", e.g. "_filt.hf.fq". Default=".fastq"}

\item{mtthread}{(Optional) Boolean, enables multithreading (not recommended in Rstudio). Default=F}

\item{trim_read_length}{(Optional) max. length at which to truncate reads. Default = 0 (no truncating)}

\item{trim5end}{(Optional) trim n reads from the start of reads (unidentified adapters, barcodes). Default=0 (no trimming)}
}
\description{
Convenience wrapper for DADA2 read processing functions. Given a directory of raw fastq files, performs platform-specific trimming and QC, writes reads to output directory.
}
\examples{
prep_454()
}
\keyword{dada2}
\keyword{processing}
\keyword{read}
